blogs can only be deleted by user that added it

+++ part4/blog/controllers/blogRouter.js
@@ -51,8 +51,22 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async(request, response) => {
-	await Blog.findByIdAndDelete(request.params.id)
-	response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token invalid'})
+	}
+	let userid = decodedToken.id
+
+	let blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() === userid.toString()){
+		await Blog.findByIdAndDelete(request.params.id)
+		return response.status(204).end()
+	} else {
+		return response.status(401).json({ error: `cannot delete other's blogs`})
+	}
+	
+	
 })
